Refactors index.ts, remove views/ OAuth workflow
After looking at the brief again, I decided not to continue
with the view-based approach since it feels like the OAuth workflow
is outside the scope of the brief, but I could be wrong -- in which
case, oops.

Reworks the index.ts file to reflect that change, as well as
changing the export to an individual instance.